+++ R/session.R
@@ -709,4 +709,3 @@ format_trace <- function(trace, show_args = FALSE) {
 sessions_dir_local <- function(cwd = getwd()) {
     file.path(cwd, ".corteza", "sessions")
 }
-